requirements.txt
flask==2.3.2
python-pptx==0.6.21
jsonschema==4.17.3
flask-limiter==2.8.1
cachetools==5.3.0
python-dotenv==1.0.0
openai>=0.27.0
requests>=2.28.0
Pillow>=9.0.0